backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
PyJWT
yoyo-migrations
